fix(failure-location): report line in test file, not test harness
Was showing the function name, file and line number of the retap module
as the location of some failures, instead of the place that the check
was initiated in the test file.
This required changing all asynch code to synchronous code so that stack
traces would not lose the stack frames in the test file.
This is still not ideal since it does not indicate the specific lines of
the JSX in the text file that is being compared, but that likely requires
some work with source maps.

closes #9